+++ src/costMatrix/sql/astarCostMatrix.sql
@@ -32,8 +32,8 @@ CREATE OR REPLACE FUNCTION pgr_astarCostMatrix(
     vids ANYARRAY,
     directed BOOLEAN DEFAULT true,
     heuristic INTEGER DEFAULT 5,
-    factor NUMERIC DEFAULT 1.0,
-    epsilon NUMERIC DEFAULT 1.0,
+    factor FLOAT DEFAULT 1.0,
+    epsilon FLOAT DEFAULT 1.0,
     OUT start_vid BIGINT,
     OUT end_vid BIGINT,
     OUT agg_cost FLOAT)
